+++ HISTORY.md
@@ -4,6 +4,15 @@ Changelog
 
 (unreleased)
 ------------
+- Feat #4: remove padx and pady. [Paul Magister]
+- Feat #4: Add map visualization: make window sizes resizable. [Paul
+  Magister]
+- Feat #4: Add map visualization and metadata info box #4. [Paul
+  Magister]
+
+
+1.0.2 (2025-06-23)
+------------------
 - Feat #3: update libs: fix example config.yaml generation. [Paul
   Magister]
 - Feat #3: Update makefile with pytree and update_funding.py. [Paul